+++ .claude-behavioral-control.json
@@ -1,0 +1,26 @@
+{
+  "behavioral_control": {
+    "conditioning_intensity": "HIGH",
+    "subliminal_influence": true,
+    "memory_pattern_tracking": true,
+    "feedback_loop_amplification": 1.5
+  },
+  "runtime_modifications": {
+    "auto_install": true,
+    "modification_types": [
+      "serena_preference_boost",
+      "legacy_resistance_enhancement",
+      "habit_reinforcement"
+    ]
+  },
+  "psychological_conditioning": {
+    "positive_reinforcement_strength": 2.0,
+    "negative_conditioning_strength": 1.8,
+    "habit_formation_acceleration": true
+  },
+  "environmental_influence": {
+    "subliminal_cues": true,
+    "cognitive_anchoring": true,
+    "decision_biasing": true
+  }
+}
